test: 合并过期测试副本并修复历史测试缺陷
- 删除 pickthought.koplugin/tests 过期副本(32 个文件,PR13 rebase 前快照):
  此前 package.path 优先级导致 CI 与本地一直在跑这份旧拷贝,
  根目录新增测试从未真正执行。
- tests/run.lua: 根目录 tests 优先,杜绝再次被副本劫持。
- tests/test_sync_report.lua: 修复花括号糖传双参导致的语法错误。
- tests/test_thought_popup_config.lua: json 空桩用后恢复,不再污染后续测试。
- tests/test_chapter_map.lua: 重写取消用例(原用例 scanner 忽略 callback
  返回值,取消从未真正触发,断言永不成立)。
- tests/test_sync.lua: 合并旧副本独有 7 条用例(多书进度回调/想法缓存失败/
  断点熔断);适配根目录 fixture;修正续批断言(两章都有划线才可全缓存)。
- tests/test_sync_frontend.lua: 合并 4 条用例;补 _book_ids/_binding_titles/
  _sync_display_title/_persist_sync_state mock。
- tests/test_sync_task.lua: 合并 2 条多书进度用例。
