--- pickthought.koplugin/tests/run.lua
@@ -1,67 +1,0 @@
--- 用法:luajit tests/run.lua(在仓库根目录)
-io.stdout:setvbuf("no") -- 非 TTY(管道/重定向)下保证汇总即时刷新
-local T = {passed = 0, failed = 0}
-function T.case(name, fn)
-    local ok, err = xpcall(fn, debug.traceback)
-    if ok then
-        T.passed = T.passed + 1
-    else
-        T.failed = T.failed + 1
-        print("FAIL: " .. name .. "\n  " .. tostring(err))
-    end
-end
-function T.ok(cond, label)
-    if not cond then error(tostring(label or "assertion failed"), 2) end
-end
-function T.eq(got, want, label)
-    if got ~= want then
-        error(string.format("%s\n  got:  %s\n  want: %s",
-            tostring(label or "not equal"), tostring(got), tostring(want)), 2)
-    end
-end
-_G.T = T
-_G.STUBS = require("tests.stubs")
-
-local files = {
-    "tests.test_smoke",
-    "tests.test_epub_reader",
-    "tests.test_annotation_style",
-    "tests.test_epub_inject",
-    "tests.test_chapter_map",
-    "tests.test_sync_budget",
-    "tests.test_sync_state",
-    "tests.test_sync_report",
-    "tests.test_batch_sync",
-    "tests.test_power_inhibit",
-    "tests.test_sync_task",
-    "tests.test_sync_gate",
-    "tests.test_thought_db",
-    "tests.test_thoughts",
-    "tests.test_annotation_compat",
-    "tests.test_thought_db_integrity",
-    "tests.test_thoughts_lru",
-    "tests.test_updater",
-    "tests.test_performance_mode",
-    "tests.test_sync",
-    "tests.test_binding",
-    "tests.test_spine_cache",
-    "tests.test_web_fetch",
-    "tests.test_sync_clean_source",
-    "tests.test_sync_frontend",
-    -- 想法弹窗重做使用大量 KOReader UI mock，放在最后避免污染其他模块。
-    "tests.test_thought_popup",
-    "tests.test_thought_popup_core",
-    "tests.test_thought_popup_pages",
-    "tests.test_thought_popup_widgets",
-    "tests.test_thought_popup_config",
-  }
-for _, name in ipairs(files) do
-    local ok, err = pcall(require, name)
-    if not ok and not tostring(err):find("module '" .. name .. "' not found", 1, true) then
-        T.failed = T.failed + 1
-        print("FAIL(load): " .. name .. "\n  " .. tostring(err))
-    end
-end
-
-print(string.format("passed=%d failed=%d", T.passed, T.failed))
-os.exit(T.failed == 0 and 0 or 1)
